requirements.txt
googletrans==4.0.0rc1
django-cors-headers==4.7.0
djangorestframework==3.15.2
